+++ api/node/VMNetInfo.go
@@ -5,10 +5,10 @@ import (
 )
 
 type VMNetInfo struct {
-	ReceivedPackets       int `json:"receivedPackets" validate:"nonzero"`
-	ReceivedThroughput    int `json:"receivedThroughput" validate:"nonzero"`
-	TransmittedPackets    int `json:"transmittedPackets" validate:"nonzero"`
-	TransmittedThroughput int `json:"transmittedThroughput" validate:"nonzero"`
+	ReceivedPackets       uint64 `json:"receivedPackets" validate:"nonzero"`
+	ReceivedThroughput    uint64 `json:"receivedThroughput" validate:"nonzero"`
+	TransmittedPackets    uint64 `json:"transmittedPackets" validate:"nonzero"`
+	TransmittedThroughput uint64 `json:"transmittedThroughput" validate:"nonzero"`
 }
 
 func (s VMNetInfo) Validate() error {
